Add auto fan if bed is warmer than threshold.
This keeps a fan running if the bed is above a certain threshold.
This can be useful to signal the user that the bed is still too warm (to
touch or to remove the built objects) or because the fan might actually
help cooling the bed.

To make it work, define the pin of the fan that should be enabled
BED_AUTO_FAN_PIN and the bed threshold temperature
BED_AUTO_FAN_TEMPERATURE.