Implement G60/G61 (pos store/restore) from Marlin2.
Storing a position, moving somewhere else and then being able to return
can be quite useful, e.g. when changing extruders and wanting to dump
some primed filament at some definted place.

Marlin-2 has the ability to store the current position with G60 and
return to it later with G61. Backport the feature to Marlin-1.1.x.

Define STORED_POSITIONS (with the number of wanted slots) to enable
the feature.