Allow defining a factor for slower extrusions.
The max speed for extrusion may not be the same as for retraction.
The work for extrusion is much harder, as the filament can only melt
with a finite speed and the pressure goes up quickly if the melting is
too slow. However, in std Marlin, the limit (set with M203 Exxx) for
retraction and extrusion is the same. Allow defining a factor
DOWNSCALE_EXTRUSION_FACTOR that is applied to the (retraction) limit
set with M203 Exxx for limiting extrusion speeds.